hi luke. moves routes to controllers, creates 404 view
