+++ controllers/manager_controller.js
@@ -1,0 +1,1 @@
+// routes for the manager page and queries for data
