+++ views/404.handlebars
@@ -1,0 +1,3 @@
+<div class="container">
+<h3 class="primary text-center">404: {{msg}}.</h3>
+</div>
